[REF] Migrates functions to pypmi.datasets.utils
Some functions shared between pypmi.datasets.loaders and fetchers are
now migrated to utils.
